Moved multiple config directories into one
