+++ .env-dev
@@ -1,5 +1,8 @@
 STEAM_API_KEY=07882958CFF10F24B804E8C4436823E2
 IP=localhost
 PORT=3000
-MONGO_URL=mongodb://localhost:27017/feathers
+MONGO_URL=mongodb://localhost:27017/tf2pickup
 SECRET=super_secret
+BETA_MODE=true
+VALIDATE_AGAINST_STEAM_GROUP=TF2PickupNETStaff
+REQUIRED_TF2_HOURS=1000
