+++ src/client/app.js
@@ -4,7 +4,7 @@ import io from 'socket.io-client';
 import hooks from 'feathers-hooks';
 import auth from 'feathers-authentication-client';
 
-import { isDev } from './config';
+import { isDev } from '../config/client';
 import configureStore from './redux/configure-store';
 
 const socket = io(window.location.origin, { path: '/ws/' });
