+++ src/client/redux/online-users/setup-listeners.js
@@ -2,10 +2,6 @@ import {
   loginUser,
   logoutUser,
 } from './actions';
-import {
-  onlineUsersLogin,
-  onlineUsersLogout,
-} from '../../../config/event-names';
 
 /**
  * Setup the event listeners for the online users.
@@ -15,11 +11,11 @@ import {
 export default function setupListeners(app) {
   const users = app.service('users');
 
-  users.on(onlineUsersLogin, (data) => {
+  users.on('online-users.login', (data) => {
     app.store.dispatch(loginUser(data));
   });
 
-  users.on(onlineUsersLogout, (data) => {
+  users.on('online-users.logout', (data) => {
     app.store.dispatch(logoutUser(data.id));
   });
 }
