+++ src/config/client.js
@@ -1,11 +1,5 @@
 export const isDev = window.location.hostname === 'localhost';
 
-export const titleSuffix = ' | TF2Pickup.net';
-
-export const sounds = {};
-
-export const storageKeys = { lastGamemode: 'lastGamemode' };
-
 export const cdnUrl = '/assets';
 
 export const imageUrl = `${cdnUrl}/images`;
