--- src/config/etf2l.js
@@ -1,9 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-
-export const divs = [
-  'N/A',
-  'Open',
-  'Mid',
-  'High',
-  'Prem',
-];
