--- src/config/event-names.js
@@ -1,4 +1,0 @@
-// Online users event names
-
-export const onlineUsersLogin = 'online-users.login';
-export const onlineUsersLogout = 'online-users.logout';
