+++ src/config/index.js
@@ -4,5 +4,3 @@ export const displayGamemodeDivs = [
   '6v6',
   '9v9',
 ];
-
-export const defaultRating = 1000;
