--- src/config/steam.js
@@ -1,7 +1,0 @@
-// Steam Group to validate users against
-export const validateUsersAgainstSteamGroup = 'TF2PickupNETStaff';
-// On first login invite user to steam group
-export const inviteToSteamGroup = '103582791435021680';
-
-// Min. required hours in TF2
-export const requiredInGameHours = 1000;
